Add version of latice returning pointers to foreign data
This is the start of providing versions of these funcs that create
foreign arrays containing the data.

This is because, as nice as the lisp versions may be, they are too
slow in many cases.

Also this means that projects with their own abstractions over local &
gpu data (like CEPL) can use the foreign versions and take ownership
of the pointers.